make onReady use zero's ready event
onReady is unlike the other events because it needs to trigger on all
listeners, rather than the event delegation approach.  It also needs
to trigger on mount if 'ready' has already happened.  This is similar
to jQuery(document).ready(...) vs the native api.

This solves the problem mentioned in #10 in a more robust way.